+++ pandapower/create.py
@@ -149,6 +149,16 @@ def create_empty_network(name="", f_hz=50., sn_mva=1, add_stdtypes=True):
                   ("step", "u4"),
                   ("max_step", "u4"),
                   ("in_service", "bool")],
+        "svc":   [("name", dtype(object)),
+                  ("bus", "u4"),
+                  ("x_l_ohm", "f8"),
+                  ("x_cvar_ohm", "f8"),
+                  ("set_vm_pu", "f8"),
+                  ("thyristor_firing_angle_degree", "f8"),
+                  ("controllable", "bool"),
+                  ("in_service", "bool"),
+                  ("min_angle_degree", "f8"),
+                  ("max_angle_degree", "f8")],
         "ext_grid": [("name", dtype(object)),
                      ("bus", "u4"),
                      ("vm_pu", "f8"),
@@ -231,6 +241,15 @@ def create_empty_network(name="", f_hz=50., sn_mva=1, add_stdtypes=True):
                       ("xtf_pu", "f8"),
                       ("sn_mva", "f8"),
                       ("in_service", 'bool')],
+        "tcsc": [("name", dtype(object)),
+                 ("from_bus", "u4"),
+                 ("to_bus", "u4"),
+                 ("x_l_ohm", "f8"),
+                 ("x_cvar_ohm", "f8"),
+                 ("set_p_to_mw", "f8"),
+                 ("thyristor_firing_angle_degree", "f8"),
+                 ("controllable", "bool"),
+                 ("in_service", "bool")],
         "dcline": [("name", dtype(object)),
                    ("from_bus", "u4"),
                    ("to_bus", "u4"),
@@ -351,6 +370,11 @@ def create_empty_network(name="", f_hz=50., sn_mva=1, add_stdtypes=True):
         "_empty_res_shunt": [("p_mw", "f8"),
                              ("q_mvar", "f8"),
                              ("vm_pu", "f8")],
+        "_empty_res_svc":   [("thyristor_firing_angle_degree", "f8"),
+                             ("x_ohm", "f8"),
+                             ("q_mvar", "f8"),
+                             ("vm_pu", "f8"),
+                             ("va_degree", "f8")],
         "_empty_res_switch": [("i_ka", "f8"),
                               ("loading_percent", "f8")],
         "_empty_res_impedance": [("p_from_mw", "f8"),
@@ -361,6 +385,19 @@ def create_empty_network(name="", f_hz=50., sn_mva=1, add_stdtypes=True):
                                  ("ql_mvar", "f8"),
                                  ("i_from_ka", "f8"),
                                  ("i_to_ka", "f8")],
+        "_empty_res_tcsc": [("thyristor_firing_angle_degree", "f8"),
+                            ("x_ohm", "f8"),
+                            ("p_from_mw", "f8"),
+                            ("q_from_mvar", "f8"),
+                            ("p_to_mw", "f8"),
+                            ("q_to_mvar", "f8"),
+                            ("pl_mw", "f8"),
+                            ("ql_mvar", "f8"),
+                            ("i_ka", "f8"),
+                            ("vm_from_pu", "f8"),
+                            ("va_from_degree", "f8"),
+                            ("vm_to_pu", "f8"),
+                            ("va_to_degree", "f8")],
         "_empty_res_dcline": [("p_from_mw", "f8"),
                               ("q_from_mvar", "f8"),
                               ("p_to_mw", "f8"),
@@ -1065,6 +1102,8 @@ def create_sgen(net, bus, p_mw, q_mvar=0, sn_mva=nan, name=None, index=None,
         **kappa (float, nan)** - the factor for the calculation of the peak short-circuit \
             current, referred to the high-voltage side (provided by the manufacturer). \
             Relevant if the generator_type is "async_doubly_fed".
+            If the superposition method is used (use_pre_fault_voltage=True), this parameter \
+            is used to pass through the max. current limit of the machine in p.u.
 
         **current_source** (bool, True) - Model this sgen as a current source during short-\
             circuit calculations; useful in some cases, for example the simulation of full-\
@@ -1095,6 +1134,8 @@ def create_sgen(net, bus, p_mw, q_mvar=0, sn_mva=nan, name=None, index=None,
                                  default_val=False, default_for_nan=True)
     if rx is not None:
         _create_column_and_set_value(net, index, rx, "rx", "sgen") # rx is always required
+    if np.isfinite(kappa):
+        _create_column_and_set_value(net, index, kappa, "kappa", "sgen")
     if generator_type is not None:
         _create_column_and_set_value(net, index, generator_type, "generator_type", "sgen", dtyp="str",
                                      default_val="current_source", default_for_nan=True)
@@ -1104,7 +1145,6 @@ def create_sgen(net, bus, p_mw, q_mvar=0, sn_mva=nan, name=None, index=None,
         _create_column_and_set_value(net, index, lrc_pu, "lrc_pu", "sgen")
     elif generator_type == "async_doubly_fed":
         _create_column_and_set_value(net, index, max_ik_ka, "max_ik_ka", "sgen")
-        _create_column_and_set_value(net, index, kappa, "kappa", "sgen")
     else:
         raise UserWarning(f"unknown sgen generator_type {generator_type}! "
                           f"Must be one of: None, 'current_source', 'async', 'async_doubly_fed'")
@@ -1218,6 +1258,8 @@ def create_sgens(net, buses, p_mw, q_mvar=0, sn_mva=nan, name=None, index=None,
                            default_val=False)
     if rx is not None:
         _add_series_to_entries(entries, index, "rx", rx)  # rx is always required
+    if np.isfinite(kappa):
+        _add_series_to_entries(entries, index, "kappa", kappa)  # is used for Type C also as a max. current limit
     if generator_type is not None:
         _add_series_to_entries(entries, index, "generator_type", generator_type, dtyp="str",
                                      default_val="current_source")
@@ -1227,7 +1269,6 @@ def create_sgens(net, buses, p_mw, q_mvar=0, sn_mva=nan, name=None, index=None,
         _add_series_to_entries(entries, index, "lrc_pu", lrc_pu)
     elif generator_type == "async_doubly_fed":
         _add_series_to_entries(entries, index, "max_ik_ka", max_ik_ka)
-        _add_series_to_entries(entries, index, "kappa", kappa)
     else:
         raise UserWarning(f"unknown sgen generator_type {generator_type}! "
                           f"Must be one of: None, 'current_source', 'async', 'async_doubly_fed'")
@@ -3697,6 +3738,63 @@ def create_shunt_as_capacitor(net, bus, q_mvar, loss_factor, **kwargs):
     return create_shunt(net, bus, q_mvar=q_mvar, p_mw=p_mw, **kwargs)
 
 
+def create_svc(net, bus, x_l_ohm, x_cvar_ohm, set_vm_pu, thyristor_firing_angle_degree,
+                name=None, controllable=True, in_service=True, index=None,
+                min_angle_degree=90, max_angle_degree=180, **kwargs):
+    """
+    Creates an SVC element - a shunt element with adjustable impedance used to control the voltage at the connected bus
+
+    Does not work if connected to "PV" bus (gen bus, ext_grid bus)
+
+    min_angle_degree, max_angle_degree are placehowlders (ignored in the Newton-Raphson power flow at the moment).
+
+    INPUT:
+        **net** (pandapowerNet) - The pandapower network in which the element is created
+
+        **bus** (int) - connection bus of the svc
+
+        **x_l_ohm** (float) - impedance of the reactor component of svc
+
+        **x_cvar_ohm** (float) - impedance of the fixed capacitor component of svc
+
+        **set_vm_pu** (float) - set-point for the bus voltage magnitude at the connection bus
+
+        **thyristor_firing_angle_degree** (float) - the value of thyristor firing angle of svc (is used directly if
+            controllable==False, otherwise is the starting point in the Newton-Raphson calculation)
+
+    OPTIONAL:
+        **name** (list of strs, None) - element name
+
+        **controllable** (bool, True) - whether the element is considered as actively controlling or
+            as a fixed shunt impedance
+
+        **in_service** (bool, True) - True for in_service or False for out of service
+
+        **index** (int, None) - Force a specified ID if it is available. If None, the
+            index one higher than the highest already existing index is selected.
+
+        **min_angle_degree** (float, 90) - minimum value of the thyristor_firing_angle_degree
+
+        **max_angle_degree** (float, 180) - maximum value of the thyristor_firing_angle_degree
+
+    OUTPUT:
+        **index** (int) - The unique ID of the created svc
+
+    """
+
+    _check_node_element(net, bus)
+
+    index = _get_index_with_check(net, "svc", index)
+
+    entries = dict(zip(["name", "bus", "x_l_ohm", "x_cvar_ohm", "set_vm_pu", "thyristor_firing_angle_degree",
+                        "controllable", "in_service", "min_angle_degree", "max_angle_degree"],
+                       [name, bus, x_l_ohm, x_cvar_ohm, set_vm_pu, thyristor_firing_angle_degree,
+                        controllable, in_service, min_angle_degree, max_angle_degree]))
+    _set_entries(net, "svc", index, **entries, **kwargs)
+
+    return index
+
+
 def create_impedance(net, from_bus, to_bus, rft_pu, xft_pu, sn_mva, rtf_pu=None, xtf_pu=None,
                      name=None, in_service=True, index=None,
                      rft0_pu=None, xft0_pu=None, rtf0_pu=None, xtf0_pu=None, **kwargs):
@@ -3748,6 +3846,72 @@ def create_impedance(net, from_bus, to_bus, rft_pu, xft_pu, sn_mva, rtf_pu=None,
         _create_column_and_set_value(net, index, xft0_pu, "xft0_pu", "impedance")
         _create_column_and_set_value(net, index, rtf0_pu, "rtf0_pu", "impedance")
         _create_column_and_set_value(net, index, xtf0_pu, "xtf0_pu", "impedance")
+
+    return index
+
+
+def create_tcsc(net, from_bus, to_bus, x_l_ohm, x_cvar_ohm, set_p_to_mw, thyristor_firing_angle_degree,
+                name=None, controllable=True, in_service=True, index=None,
+                min_angle_degree=90, max_angle_degree=180, **kwargs):
+    """
+    Creates a TCSC element - series impedance compensator to control series reactance.
+    The TCSC device allows controlling the active power flow throgh the path it is connected in.
+
+    Multiple TCSC elements in net are possible.
+    Unfortunately, TCSC is not implemented for the case when multiple TCSC elements
+    have the same from_bus or the same to_bus.
+
+    Note: in the Newton-Raphson power flow calculation, the initial voltage vector is adjusted slightly
+    if the initial voltage at the from bus is the same as at the to_bus to avoid
+    some terms in J (for TCSC) becoming zero.
+
+    min_angle_degree, max_angle_degree are placehowlders (ignored in the Newton-Raphson power flow at the moment).
+
+    INPUT:
+        **net** (pandapowerNet) - The pandapower network in which the element is created
+
+        **from_bus** (int) - starting bus of the tcsc
+
+        **to_bus** (int) - ending bus of the tcsc
+
+        **x_l_ohm** (float) - impedance of the reactor component of tcsc
+
+        **x_cvar_ohm** (float) - impedance of the fixed capacitor component of tcsc
+
+        **set_p_to_mw** (float) - set-point for the branch active power at the to_bus
+
+        **thyristor_firing_angle_degree** (float) - the value of thyristor firing angle of tcsc (is used directly if
+            controllable==False, otherwise is the starting point in the Newton-Raphson calculation)
+
+    OPTIONAL:
+        **name** (list of strs, None) - element name
+
+        **controllable** (bool, True) - whether the element is considered as actively controlling
+            or as a fixed series impedance
+
+        **in_service** (bool, True) - True for in_service or False for out of service
+
+        **index** (int, None) - Force a specified ID if it is available. If None, the
+            index one higher than the highest already existing index is selected.
+
+        **min_angle_degree** (float, 90) - minimum value of the thyristor_firing_angle_degree
+
+        **max_angle_degree** (float, 180) - maximum value of the thyristor_firing_angle_degree
+
+    OUTPUT:
+        **index** (int) - The unique ID of the created tcsc
+
+    """
+    index = _get_index_with_check(net, "tcsc", index)
+
+    _check_branch_element(net, "TCSC", index, from_bus, to_bus)
+
+    columns = ["name", "from_bus", "to_bus", "x_l_ohm", "x_cvar_ohm", "set_p_to_mw", "thyristor_firing_angle_degree",
+               "controllable", "in_service", "min_angle_degree", "max_angle_degree"]
+    values = [name, from_bus, to_bus, x_l_ohm, x_cvar_ohm, set_p_to_mw, thyristor_firing_angle_degree,
+              controllable, in_service, min_angle_degree, max_angle_degree]
+    entries = dict(zip(columns, values))
+    _set_entries(net, "tcsc", index, **entries, **kwargs)
 
     return index